requirements.txt
Flask==3.0.2
google-cloud-storage==1.42.0
google-cloud-speech==2.10.0
openai==0.10.2
google-cloud-secret-manager==2.7.0
gunicorn==20.0.4
Werkzeug>=3.0.1
